# shared/api/gateway/v1/gateway.openapi.yaml
openapi: 3.0.3
info:
  title: "Gateway API"
  version: 1.0.0
  description: "API для работы с менеджером задач"
x-ogen:
  target: ./shared/pkg/openapi/gateway/v1
  package: gateway_v1
  clean: true

tags:
  - name: Board
    description: "Операции с досками"
  - name: Column
    description: "Операции с колонками"

paths:
  /api/v1/boards:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoardRequestBody"
      summary: "Создание доски"
      operationId: BoardCreate
      tags:
        - Board
      responses:
        '200':
          description: "Доска успешно создана"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBoardResponse"
        '400':
          description: "Ошибка запроса"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        '500':
          description: "Ошибка сервера"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        default:
          description: "Прочие ошибки"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
    get:
      summary: "Получение досок"
      operationId: BoardGetAll
      tags:
        - Board
      responses:
        '200':
          description: "Успешное получение досок"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllBoardResponse"
        '400':
          description: "Ошибка запроса"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        default:
          description: "Прочие ошибки"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
  /api/v1/boards/{id}/columns:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID доски
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateColumnRequestBody"
      summary: "Создание колонки"
      operationId: ColumnCreate
      tags:
        - Column
      responses:
        '200':
          description: "Колонка успешно создана"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateColumnResponse"
        '400':
          description: "Ошибка запроса"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        '500':
          description: "Ошибка сервера"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        default:
          description: "Прочие ошибки"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
  /api/v1/boards/{boardId}/columns/{columnId}:
    delete:
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: integer
          description: ID доски
        - in: path
          name: columnId
          required: true
          schema:
            type: integer
          description: ID колонки
      summary: "Удаление колонки"
      operationId: ColumnDelete
      tags:
        - Column
      responses:
        '204':
          description: "Колонка успешно удалена"
        '400':
          description: "Ошибка запроса"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        '500':
          description: "Ошибка сервера"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        default:
          description: "Прочие ошибки"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
  /api/v1/boards/{boardId}/columns/{columnId}/update-position:
    patch:
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: integer
          description: ID доски
        - in: path
          name: columnId
          required: true
          schema:
            type: integer
          description: ID колонки
      summary: "Изменение позиции колонки"
      operationId: ColumnUpdatePosition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColumnUpdatePositionRequestBody"
      tags:
        - Column
      responses:
        '204':
          description: "Колонка успешно удалена"
        '400':
          description: "Ошибка запроса"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        '500':
          description: "Ошибка сервера"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        default:
          description: "Прочие ошибки"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
  /api/v1/boards/{boardId}/columns/{columnId}/tasks:
    post:
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: integer
          description: ID доски
        - in: path
          name: columnId
          required: true
          schema:
            type: integer
          description: ID колонки
      summary: "Создание задачи"
      operationId: TaskCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequestBody"
      tags:
        - Column
      responses:
        '200':
          description: "Задача успешно создана"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCreateResponse"
        '400':
          description: "Ошибка запроса"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        '500':
          description: "Ошибка сервера"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
        default:
          description: "Прочие ошибки"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericError"
components:
  schemas:
    CreateBoardRequestBody:
      type: object
      required:
        - name
        - description
        - owner_id
      properties:
        name:
          type: string
          description: "Название доски"
          example: "Project"
        description:
          type: string
          description: "Описание доски"
          example: "Запланированное"
        owner_id:
          type: integer
          description: "ID автора доски"
          example: 1
    CreateColumnRequestBody:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: "Название колонки"
          example: "TODO"
    ColumnUpdatePositionRequestBody:
      type: object
      required:
        - left_position
        - right_position
      properties:
        left_position:
          type: number
          description: "Номер позиции слева"
          example: 2000
        right_position:
          type: number
          description: "Номер позиции справа"
          example: 3000
    TaskCreateRequestBody:
      type: object
      required:
        - title
        - description
        - assignee
      properties:
        title:
          type: string
          description: "Название задачи"
          example: "Task"
        description:
          type: string
          description: "Описание задачи"
          example: "Description"
        assignee:
          type: string
          description: "Ответственный по задаче"
          example: "Assignee"
    CreateBoardResponse:
      type: object
      required:
        - board_id
      properties:
        board_id:
          type: integer
          description: "ID созданной доски"
          example: 1
    CreateColumnResponse:
      type: object
      required:
        - column_id
      properties:
        column_id:
          type: integer
          description: "ID созданной колонки"
          example: 1
    TaskCreateResponse:
      type: object
      required:
        - task_id
      properties:
        task_id:
          type: integer
          description: "ID созданной задачи"
          example: 1
    GetAllBoardResponse:
      type: array
      items:
        type: object
        required:
          - id
          - name
          - description
        properties:
          id:
            type: integer
            description: "ID доски"
            example: 1
          name:
            type: string
            description: "Название доски"
            example: "Test name"
          description:
            type: string
            description: "Описание доски"
            example: "Test description"
    BadRequestError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: "HTTP-код ошибки"
          example: 400
        message:
          type: string
          description: "Описание ошибки"
          example: "Ошибка запроса"
    InternalServerError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: "HTTP-код ошибки"
          example: 500
        message:
          type: string
          description: "Описание ошибки"
          example: "Ошибка сервера"
    GenericError:
      type: object
      properties:
        code:
          type: integer
          description: "HTTP-код ошибки"
        message:
          type: string
          description: "Описание ошибки"